+++ HISTORY.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### New Features
 * Add an option `strict_bytes_per_sync` that causes a file-writing thread to block rather than exceed the limit on bytes pending writeback specified by `bytes_per_sync` or `wal_bytes_per_sync`.
+* When reading from option file/string/map, customized envs can be filled according to object registry.
 
 # 5.17.2 (10/24/2018)
 ### Bug Fixes